clean up results props
